pyproject.toml
[tool.poetry]
name = "repl_python3_beta_Sexy-Guru-bot"
version = "0.1.0"
description = ""
authors = ["BrunoGrifo <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
jokepy = "^1.0.3"
Flask = "^2.0.2"
requests-oauthlib = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
